Fix error caused by use of same Grab in multiple threads
That's an error I encountered: https://github.com/lorien/grab/issues/185
And the Grab author says that the problem is caused by the reuse of the
same Grab instance in multiple threads. Which is indeed the case we
have here.

Also, looking back, with some long server response times there might be
more bugs caused by one thread's parser modifying self.g.doc while
another thread is doing something with it.

So, it makes total sense to have per-thread instances of Grab.

To avoid the need to rewrite all the code I add @property self.g that
either returns existing Grab instance for this thread, or creates and
initialized a new one.

But we have to patch Grab instances so they use the same Cookie storage.
Kinda like different windows of the browser. This does not seem like a
very supported trick, and it will break if we use some advanced Grab
functionality like .adopt or something. But so fat it seems to be a
pretty good solution.